Extract Android analytics identity linking
