+++ apps/android/data/local/src/androidTest/java/com/flashcardsopensourceapp/data/local/repository/cloudsync/guest/CloudGuestSessionCoordinatorTest.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.flashcardsopensourceapp.data.local.ai.remote.GuestCloudSessionCreator
 import com.flashcardsopensourceapp.data.local.cloud.PendingGuestUpgradeState
-import com.flashcardsopensourceapp.data.local.cloud.remote.CloudRemoteException
 import com.flashcardsopensourceapp.data.local.database.entities.SyncStateEntity
 import com.flashcardsopensourceapp.data.local.model.ai.StoredGuestAiSession
 import com.flashcardsopensourceapp.data.local.model.cloud.CloudAccountState
@@ -23,11 +22,9 @@ import com.flashcardsopensourceapp.data.local.repository.cloudsync.support.creat
 import com.flashcardsopensourceapp.data.local.repository.cloudsync.support.createCloudWorkspaceSummary
 import com.flashcardsopensourceapp.data.local.repository.cloudsync.support.createStoredCloudCredentials
 import com.flashcardsopensourceapp.data.local.repository.cloudsync.support.createStoredGuestAiSession
-import com.flashcardsopensourceapp.data.local.repository.cloudsync.support.createSyncCardOutboxEntry
 import java.net.SocketTimeoutException
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
-import org.json.JSONObject
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
@@ -81,7 +78,7 @@ class CloudGuestSessionCoordinatorTest {
 
     @Test
     fun linkedStateWithMissingCredentialsMarksRecoveryAndPreservesLocalData() = runBlocking {
-        val preservationState = seedCredentialRecoveryLocalData()
+        val preservationState = seedCredentialRecoveryLocalData(environment = environment)
         val coordinator = environment.createCloudGuestSessionCoordinator(
             remoteGateway = FakeCloudRemoteGateway.standard()
         )
@@ -107,7 +104,10 @@ class CloudGuestSessionCoordinatorTest {
         assertEquals("user@example.com", recoveryState.linkedEmail)
         assertEquals(CloudServiceConfigurationMode.OFFICIAL, recoveryState.configurationMode)
         assertEquals("https://api.flashcards-open-source-app.com/v1", recoveryState.apiBaseUrl)
-        assertCredentialRecoveryPreservedLocalData(preservationState = preservationState)
+        assertCredentialRecoveryPreservedLocalData(
+            environment = environment,
+            preservationState = preservationState
+        )
         assertEquals(CloudAccountState.DISCONNECTED, environment.cloudPreferencesStore.currentCloudSettings().cloudState)
         assertEquals(preservationState.installationId, environment.cloudPreferencesStore.currentCloudSettings().installationId)
         assertEquals(preservationState.workspaceId, environment.cloudPreferencesStore.currentCloudSettings().activeWorkspaceId)
@@ -115,7 +115,7 @@ class CloudGuestSessionCoordinatorTest {
 
     @Test
     fun linkedStateWithInvalidActiveWorkspaceAndMissingCredentialsMarksRecovery() = runBlocking {
-        val preservationState = seedCredentialRecoveryLocalData()
+        val preservationState = seedCredentialRecoveryLocalData(environment = environment)
         val coordinator = environment.createCloudGuestSessionCoordinator(
             remoteGateway = FakeCloudRemoteGateway.standard()
         )
@@ -135,14 +135,17 @@ class CloudGuestSessionCoordinatorTest {
         assertEquals(CloudCredentialRecoveryReason.LINKED_CREDENTIALS_MISSING, recoveryState.reason)
         assertEquals(preservationState.workspaceId, recoveryState.linkedWorkspaceId)
         assertEquals(preservationState.workspaceId, recoveryState.activeWorkspaceId)
-        assertCredentialRecoveryPreservedLocalData(preservationState = preservationState)
+        assertCredentialRecoveryPreservedLocalData(
+            environment = environment,
+            preservationState = preservationState
+        )
         assertEquals(CloudAccountState.DISCONNECTED, environment.cloudPreferencesStore.currentCloudSettings().cloudState)
         assertEquals(preservationState.workspaceId, environment.cloudPreferencesStore.currentCloudSettings().activeWorkspaceId)
     }
 
     @Test
     fun guestStateWithMissingSessionMarksRecoveryAndPreservesLocalData() = runBlocking {
-        val preservationState = seedCredentialRecoveryLocalData()
+        val preservationState = seedCredentialRecoveryLocalData(environment = environment)
         val coordinator = environment.createCloudGuestSessionCoordinator(
             remoteGateway = FakeCloudRemoteGateway.standard()
         )
@@ -166,7 +169,10 @@ class CloudGuestSessionCoordinatorTest {
         assertEquals(preservationState.workspaceId, recoveryState.linkedWorkspaceId)
         assertEquals(preservationState.workspaceId, recoveryState.activeWorkspaceId)
         assertNull(recoveryState.linkedEmail)
-        assertCredentialRecoveryPreservedLocalData(preservationState = preservationState)
+        assertCredentialRecoveryPreservedLocalData(
+            environment = environment,
+            preservationState = preservationState
+        )
         assertEquals(CloudAccountState.DISCONNECTED, environment.cloudPreferencesStore.currentCloudSettings().cloudState)
         assertEquals(preservationState.installationId, environment.cloudPreferencesStore.currentCloudSettings().installationId)
         assertEquals(preservationState.workspaceId, environment.cloudPreferencesStore.currentCloudSettings().activeWorkspaceId)
@@ -174,7 +180,7 @@ class CloudGuestSessionCoordinatorTest {
 
     @Test
     fun guestStateWithInvalidActiveWorkspaceAndMissingSessionMarksRecovery() = runBlocking {
-        val preservationState = seedCredentialRecoveryLocalData()
+        val preservationState = seedCredentialRecoveryLocalData(environment = environment)
         val coordinator = environment.createCloudGuestSessionCoordinator(
             remoteGateway = FakeCloudRemoteGateway.standard()
         )
@@ -194,7 +200,10 @@ class CloudGuestSessionCoordinatorTest {
         assertEquals(CloudCredentialRecoveryReason.GUEST_SESSION_MISSING, recoveryState.reason)
         assertEquals(preservationState.workspaceId, recoveryState.linkedWorkspaceId)
         assertEquals(preservationState.workspaceId, recoveryState.activeWorkspaceId)
-        assertCredentialRecoveryPreservedLocalData(preservationState = preservationState)
+        assertCredentialRecoveryPreservedLocalData(
+            environment = environment,
+            preservationState = preservationState
+        )
         assertEquals(CloudAccountState.DISCONNECTED, environment.cloudPreferencesStore.currentCloudSettings().cloudState)
         assertEquals(preservationState.workspaceId, environment.cloudPreferencesStore.currentCloudSettings().activeWorkspaceId)
     }
@@ -746,220 +755,6 @@ class CloudGuestSessionCoordinatorTest {
         )
     }
 
-    @Test
-    fun analyticsGuestIdentityLinkSendsStoredGuestTokenAndClearsStoredSessions() = runBlocking {
-        val localWorkspaceId = environment.requireLocalWorkspaceId()
-        val remoteGateway = FakeCloudRemoteGateway.standard()
-        val coordinator = environment.createCloudGuestSessionCoordinator(remoteGateway = remoteGateway)
-        environment.prepareLinkedCloudIdentity(localWorkspaceId = localWorkspaceId)
-        storeAnalyticsOnlyGuestSession(guestToken = "analytics-guest-token")
-
-        coordinator.linkAnalyticsGuestIdentityToSignedInAccount()
-
-        assertEquals(listOf("analytics-guest-token"), remoteGateway.linkGuestIdentityGuestTokens)
-        // The account's identity row is written by the first request that loads a request context,
-        // so exactly one has to precede the link or it earns `409 ACCOUNT_REQUIRED`.
-        assertEquals(1, remoteGateway.fetchCloudAccountCalls)
-        assertNull(
-            environment.guestAiSessionStore.loadAnySession(
-                configuration = makeOfficialCloudServiceConfiguration()
-            )
-        )
-    }
-
-    @Test
-    fun analyticsGuestIdentityLinkSkipsGuestSessionThatOwnsCloudData() = runBlocking {
-        val localWorkspaceId = environment.requireLocalWorkspaceId()
-        val remoteGateway = FakeCloudRemoteGateway.standard()
-        val coordinator = environment.createCloudGuestSessionCoordinator(remoteGateway = remoteGateway)
-        environment.prepareLinkedCloudIdentity(localWorkspaceId = localWorkspaceId)
-        environment.guestAiSessionStore.saveSession(
-            localWorkspaceId = null,
-            session = createStoredGuestAiSession(
-                workspaceId = "guest-workspace",
-                configurationMode = CloudServiceConfigurationMode.OFFICIAL,
-                apiBaseUrl = "https://api.flashcards-open-source-app.com/v1",
-                guestToken = "cloud-guest-token",
-                userId = "cloud-guest-user"
-            )
-        )
-
-        coordinator.linkAnalyticsGuestIdentityToSignedInAccount()
-
-        // That guest converts through `/guest-auth/upgrade/complete`, which writes the same identity
-        // link; this route would revoke the session that flow still needs.
-        assertTrue(remoteGateway.linkGuestIdentityGuestTokens.isEmpty())
-        assertEquals(
-            "cloud-guest-token",
-            environment.guestAiSessionStore.loadAnySession(
-                configuration = makeOfficialCloudServiceConfiguration()
-            )?.guestToken
-        )
-    }
-
-    @Test
-    fun analyticsGuestIdentityLinkKeepsGuestTokenAndStopsRetryingWhenUpgradeIsRequired() = runBlocking {
-        val localWorkspaceId = environment.requireLocalWorkspaceId()
-        val remoteGateway = FakeCloudRemoteGateway.standard()
-        remoteGateway.setLinkGuestIdentityError(
-            error = createCloudRemoteError(
-                statusCode = 409,
-                errorCode = "GUEST_IDENTITY_LINK_UPGRADE_REQUIRED",
-                path = "/guest-auth/identity/link"
-            )
-        )
-        val coordinator = environment.createCloudGuestSessionCoordinator(remoteGateway = remoteGateway)
-        environment.prepareLinkedCloudIdentity(localWorkspaceId = localWorkspaceId)
-        storeAnalyticsOnlyGuestSession(guestToken = "analytics-guest-token")
-
-        coordinator.linkAnalyticsGuestIdentityToSignedInAccount()
-
-        // The token is kept: it names the owner of data only the upgrade flow can transfer.
-        val retiredSession = environment.guestAiSessionStore.loadAnySession(
-            configuration = makeOfficialCloudServiceConfiguration()
-        )
-        assertEquals("analytics-guest-token", retiredSession?.guestToken)
-        assertEquals(false, retiredSession?.isAnalyticsOnly)
-
-        coordinator.linkAnalyticsGuestIdentityToSignedInAccount()
-
-        // Correcting the marker is what stops a link that can never succeed from being retried on
-        // every app start and every sign-in.
-        assertEquals(listOf("analytics-guest-token"), remoteGateway.linkGuestIdentityGuestTokens)
-    }
-
-    @Test
-    fun analyticsGuestIdentityLinkDropsGuestTokenOwnedByAnotherAccount() = runBlocking {
-        val localWorkspaceId = environment.requireLocalWorkspaceId()
-        val remoteGateway = FakeCloudRemoteGateway.standard()
-        remoteGateway.setLinkGuestIdentityError(
-            error = createCloudRemoteError(
-                statusCode = 409,
-                errorCode = "GUEST_IDENTITY_LINK_OTHER_ACCOUNT",
-                path = "/guest-auth/identity/link"
-            )
-        )
-        val coordinator = environment.createCloudGuestSessionCoordinator(remoteGateway = remoteGateway)
-        environment.prepareLinkedCloudIdentity(localWorkspaceId = localWorkspaceId)
-        storeAnalyticsOnlyGuestSession(guestToken = "analytics-guest-token")
-
-        coordinator.linkAnalyticsGuestIdentityToSignedInAccount()
-
-        assertEquals(listOf("analytics-guest-token"), remoteGateway.linkGuestIdentityGuestTokens)
-        // Terminal: the credential is not this install's, so it must not stay as an analytics one.
-        assertNull(
-            environment.guestAiSessionStore.loadAnySession(
-                configuration = makeOfficialCloudServiceConfiguration()
-            )
-        )
-    }
-
-    @Test
-    fun analyticsGuestIdentityLinkKeepsGuestTokenWhenTheLinkFailsRetryably() = runBlocking {
-        val localWorkspaceId = environment.requireLocalWorkspaceId()
-        val remoteGateway = FakeCloudRemoteGateway.standard()
-        remoteGateway.setLinkGuestIdentityError(
-            error = createCloudRemoteError(
-                statusCode = 500,
-                errorCode = null,
-                path = "/guest-auth/identity/link"
-            )
-        )
-        val coordinator = environment.createCloudGuestSessionCoordinator(remoteGateway = remoteGateway)
-        environment.prepareLinkedCloudIdentity(localWorkspaceId = localWorkspaceId)
-        storeAnalyticsOnlyGuestSession(guestToken = "analytics-guest-token")
-
-        try {
-            coordinator.linkAnalyticsGuestIdentityToSignedInAccount()
-            throw AssertionError("Expected a retryable link failure to be rethrown.")
-        } catch (error: CloudRemoteException) {
-            assertEquals(500, error.statusCode)
-        }
-
-        // A 5xx leaves this guest's tail unclaimed, because a success is what says the link landed,
-        // so the retry is mandatory and the token has to survive for it.
-        val keptSession = environment.guestAiSessionStore.loadAnySession(
-            configuration = makeOfficialCloudServiceConfiguration()
-        )
-        assertEquals("analytics-guest-token", keptSession?.guestToken)
-        assertTrue(keptSession?.isAnalyticsOnly == true)
-    }
-
-    /**
-     * The credential must belong to the account this install is linked to. Linking on a mismatch
-     * would attribute the guest's whole pre-sign-in tail to the wrong account, and
-     * `analytics.identity_links` is first-link-wins with no repair path — while answering the
-     * mismatch the way `authenticatedSession()` does would destroy local data instead.
-     */
-    @Test
-    fun analyticsGuestIdentityLinkBailsWhenTheCredentialIsForAnotherAccount() = runBlocking {
-        val preservationState = seedCredentialRecoveryLocalData()
-        val remoteGateway = FakeCloudRemoteGateway.forAccountSnapshot(
-            accountSnapshot = createCloudAccountSnapshot(
-                userId = "user-2",
-                email = "other@example.com",
-                workspaces = listOf(
-                    createCloudWorkspaceSummary(
-                        workspaceId = preservationState.workspaceId,
-                        name = "Personal",
-                        createdAtMillis = 100L,
-                        isSelected = true
-                    )
-                )
-            )
-        )
-        val coordinator = environment.createCloudGuestSessionCoordinator(remoteGateway = remoteGateway)
-        environment.prepareLinkedCloudIdentity(localWorkspaceId = preservationState.workspaceId)
-        storeAnalyticsOnlyGuestSession(guestToken = "analytics-guest-token")
-
-        coordinator.linkAnalyticsGuestIdentityToSignedInAccount()
-
-        assertTrue(remoteGateway.linkGuestIdentityGuestTokens.isEmpty())
-        assertCredentialRecoveryPreservedLocalData(preservationState = preservationState)
-        assertEquals(
-            "analytics-guest-token",
-            environment.guestAiSessionStore.loadAnySession(
-                configuration = makeOfficialCloudServiceConfiguration()
-            )?.guestToken
-        )
-    }
-
-    /**
-     * The background link runs unattended on every app start and after every sign-in. A deleted
-     * account must therefore never reach a destructive local reset from here: sync answers that
-     * condition by preserving local data, and nothing about analytics may pre-empt it.
-     */
-    @Test
-    fun analyticsGuestIdentityLinkPreservesLocalDataWhenTheAccountIsDeleted() = runBlocking {
-        val preservationState = seedCredentialRecoveryLocalData()
-        val remoteGateway = FakeCloudRemoteGateway.forFetchAccountError(
-            fetchAccountError = createCloudRemoteError(
-                statusCode = 410,
-                errorCode = "ACCOUNT_DELETED",
-                path = "/me"
-            )
-        )
-        val coordinator = environment.createCloudGuestSessionCoordinator(remoteGateway = remoteGateway)
-        environment.prepareLinkedCloudIdentity(localWorkspaceId = preservationState.workspaceId)
-        storeAnalyticsOnlyGuestSession(guestToken = "analytics-guest-token")
-
-        coordinator.linkAnalyticsGuestIdentityToSignedInAccount()
-
-        assertTrue(remoteGateway.linkGuestIdentityGuestTokens.isEmpty())
-        assertCredentialRecoveryPreservedLocalData(preservationState = preservationState)
-        assertEquals(
-            preservationState.installationId,
-            environment.cloudPreferencesStore.currentCloudSettings().installationId
-        )
-        assertEquals(CloudAccountState.LINKED, environment.cloudPreferencesStore.currentCloudSettings().cloudState)
-        assertNotNull(environment.cloudPreferencesStore.loadCredentials())
-        assertEquals(
-            "analytics-guest-token",
-            environment.guestAiSessionStore.loadAnySession(
-                configuration = makeOfficialCloudServiceConfiguration()
-            )?.guestToken
-        )
-    }
 
     /**
      * The creation idempotency key is minted once and kept until the session is committed, so a
@@ -1017,82 +812,7 @@ class CloudGuestSessionCoordinatorTest {
         assertNull(environment.guestAiSessionStore.loadPendingCreationIdempotencyKey())
     }
 
-    private fun storeAnalyticsOnlyGuestSession(guestToken: String) {
-        environment.guestAiSessionStore.saveSession(
-            localWorkspaceId = null,
-            session = createStoredGuestAiSession(
-                workspaceId = "analytics-guest-workspace",
-                configurationMode = CloudServiceConfigurationMode.OFFICIAL,
-                apiBaseUrl = "https://api.flashcards-open-source-app.com/v1",
-                guestToken = guestToken,
-                userId = "analytics-guest-user",
-                isAnalyticsOnly = true
-            )
-        )
-    }
-
-    private fun createCloudRemoteError(
-        statusCode: Int,
-        errorCode: String?,
-        path: String
-    ): CloudRemoteException {
-        return CloudRemoteException(
-            message = "Cloud request failed with status $statusCode for $path",
-            statusCode = statusCode,
-            responseBody = JSONObject()
-                .put("code", errorCode ?: JSONObject.NULL)
-                .put("requestId", "request-1")
-                .toString(),
-            errorCode = errorCode,
-            requestId = "request-1",
-            syncConflict = null,
-            androidObservationAlreadyCaptured = false
-        )
-    }
-
-    private suspend fun seedCredentialRecoveryLocalData(): CredentialRecoveryPreservationState {
-        val workspaceId = environment.requireLocalWorkspaceId()
-        val installationId = environment.cloudPreferencesStore.currentCloudSettings().installationId
-        val cardId = environment.seedWorkspaceData(workspaceId = workspaceId)
-        val card = requireNotNull(environment.database.cardDao().loadCard(cardId = cardId)) {
-            "Expected seeded card."
-        }
-        environment.database.outboxDao().insertOutboxEntry(
-            createSyncCardOutboxEntry(
-                outboxEntryId = "outbox-recovery-$workspaceId",
-                workspaceId = workspaceId,
-                installationId = installationId,
-                card = card,
-                createdAtMillis = 300L
-            )
-        )
-        return CredentialRecoveryPreservationState(
-            workspaceId = workspaceId,
-            installationId = installationId,
-            cardId = cardId
-        )
-    }
-
-    private suspend fun assertCredentialRecoveryPreservedLocalData(
-        preservationState: CredentialRecoveryPreservationState
-    ) {
-        assertEquals(
-            preservationState.workspaceId,
-            environment.database.workspaceDao().loadAnyWorkspace()?.workspaceId
-        )
-        assertEquals(1, environment.database.workspaceDao().countWorkspaces())
-        assertEquals(1, environment.database.cardDao().loadCards(workspaceId = preservationState.workspaceId).count())
-        assertNotNull(environment.database.cardDao().loadCard(cardId = preservationState.cardId))
-        assertEquals(1, environment.database.reviewLogDao().countReviewLogs(workspaceId = preservationState.workspaceId))
-        assertEquals(1, environment.database.outboxDao().countOutboxEntries())
-    }
 }
-
-private data class CredentialRecoveryPreservationState(
-    val workspaceId: String,
-    val installationId: String,
-    val cardId: String
-)
 
 /**
  * [initialFailureCount] fails that many attempts before succeeding, standing in for a creation whose
